adding initial ROI

+++ jsk_pcl_ros/src/voxel_grid_downsample_decoder_nodelet.cpp
@@ -104,10 +104,10 @@ namespace jsk_pcl_ros
     }
     else {
       if (previous_id_ + 1 != id) { // the point cloud is not continuous
-	if (previous_id_ != pc_buffer_.size() - 1) { // if not the last one
-	  // make pc_buffer_ shorten
-	  pc_buffer_.resize(previous_id_ + 1);
-	}
+        if (previous_id_ != (int)pc_buffer_.size() - 1) { // if not the last one
+          // make pc_buffer_ shorten
+          pc_buffer_.resize(previous_id_ + 1);
+        }
       }
     }
     previous_id_ = id;          // update the previous one
